1: Counters added to prevent saying the same warning twice. Manual play mode coming soon(TM) 2: Warnings are only used if they are new or updates. They will not announce ANY cancellations. 3: Names have been simplified into "alert#.wav" instead of individual alerts. It's just easier. I do suggest saving these in a different folder if you intend not to overwrite. 4: INFINITY LOOP! It will ACTUALLY go on for as long as it runs now. This is due to a while loop function - so you have to manually stop it if you don't want it to continue announcing. 5: Instead of annoucing the entire alert, it will print the alert, and then it will annouce the alert in "Warning for Port ST Lucie, GA" format.
This is overall a much better system and relies on the run console instead of using the python editor. It makes it so much more efficient - and also allows those with 0 coding experience to boot it in python and call it a day.
It is still suggested to set your own user agent, but it is using a temporary email for now. This email is maintained exclusively for issues and suggestions.
pyoeas@gmail.com